backend/requirements.txt
Django==6.0.1
Pillow==12.1.0
django-filter==24.3
djangorestframework==3.15.2
django-cors-headers==4.6.0
